# internal/server/generated/types.gen.go
// Package generated provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.15.0 DO NOT EDIT.
package generated

// Defines values for ErrorCode.
const (
	BADREQUEST          ErrorCode = "BAD_REQUEST"
	INTERNALSERVERERROR ErrorCode = "INTERNAL_SERVER_ERROR"
)

// Error Детальная информация об ошибке
type Error struct {
	// Attribute Атрибут, с которым связана ошибка
	Attribute *string `json:"attribute,omitempty"`

	// Code Код ошибки:
	//  * `BAD_REQUEST` - Некорректный запрос к серверу
	//  * `INTERNAL_SERVER_ERROR` - Внутренняя ошибка сервера
	Code ErrorCode `json:"code"`

	// Message Описание ошибки
	Message *string `json:"message,omitempty"`
}

// ErrorCode Код ошибки:
//   - `BAD_REQUEST` - Некорректный запрос к серверу
//   - `INTERNAL_SERVER_ERROR` - Внутренняя ошибка сервера
type ErrorCode string

// RequestCreateTask Тело запроса для создания задачи
type RequestCreateTask struct {
	Description *string `json:"description,omitempty"`
	Title       string  `json:"title"`
}

// RespSuccess Всё хорошо
type RespSuccess struct {
	// Ok ok
	Ok bool `json:"ok"`
}

// ResponseCreateTask Тело ответа для создания задачи
type ResponseCreateTask struct {
	Description *string `json:"description,omitempty"`

	// Id Уникальный идентификатор
	Id    UUID   `json:"id"`
	Title string `json:"title"`
}

// ResponseGetTasks Тело ответа получения задач
type ResponseGetTasks = []Task

// Task defines model for Task.
type Task struct {
	Description *string `json:"description,omitempty"`

	// Id Уникальный идентификатор
	Id    UUID   `json:"id"`
	Title string `json:"title"`
}

// UUID Уникальный идентификатор
type UUID = string

// ID Уникальный идентификатор
type ID = UUID

// Success Всё хорошо
type Success = RespSuccess

// CreateTaskJSONRequestBody defines body for CreateTask for application/json ContentType.
type CreateTaskJSONRequestBody = RequestCreateTask